% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydrograph.R
\name{hydrograph_plot}
\alias{hydrograph_plot}
\title{Hydrograph plot}
\usage{
hydrograph_plot(
  input = matrix(ncol = 2, nrow = 2),
  streamflow = input[, 2],
  timeSeries = input[, 1],
  streamflow2 = NULL,
  precip = NULL,
  begin = 1,
  endindex = length(streamflow),
  P.units = "",
  S.units = "",
  S1.col = "black",
  S2.col = "red",
  left_label = "Streamflow",
  right_label,
  streamflow3 = NULL,
  streamflow4 = NULL,
  precip2 = NULL
)
}
\arguments{
\item{input}{This is a data frame with at least two columns of data First column: Must be a time/date series Second column: If including precip, precip. Otherwise, streamflow - AND do not include any other columns Third column: (Only include if precip is in the second column) First streamflow dataset Forth column: (optional) Second streamflow dataset}

\item{streamflow}{vector of streamflow values - not necessary if using "input"}

\item{timeSeries}{vector of time or date values - not necessary if using "input"}

\item{streamflow2}{optional vector of streamflow values - not necessary if using "input"}

\item{precip}{vector of precipitation values - not necessary if using "input"}

\item{begin}{If you don't want to graph the whole data set, include the index where you would like to start}

\item{endindex}{If you don't want to graph the whole dataset, include the index where you would like to end}

\item{P.units}{Precipitation units (character)}

\item{S.units}{Streamflow units (character). Users who have volumetric flow will have the superscript correctly formatted if they enter "m3s" or "ft3s"}

\item{S1.col}{color of the line representing streamflow}

\item{S2.col}{color of the line representing second streamflow data}

\item{left_label}{Character for left label}

\item{right_label}{Character for left label}

\item{streamflow3}{optional vector of additional streamflow values}

\item{streamflow4}{optional vector of additional streamflow values}

\item{precip2}{optional vector of a second precip gage data if you are interested in comparing precip inputs}
}
\value{
ploted data object
}
\description{
Creates a hydrograph with one or two streamflow data sets, and can include a precipitation hyetograph adapted form `EcoHydRology`
}
\note{
This function can now take NA values. If you chose to use the input argument but are not including precipitation, input must only have two columns - you will need to add the second streamflow dataset using the "streamflow2" argument.
}
\examples{
\dontrun{
png("data/test.png")
rainfall <- dataplot::dataset

hydrograph_plot(timeSeries = rainfall$`Date Time`,
           streamflow = rainfall$Discharge,
           precip = rainfall$`Rainfall depth (mm)`,
           S1.col = "red",
           left_label = "Discharge", S.units = "mm",
           right_label = "Rainfall", P.units = "mm")
dev.off()
}
}
