% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fonts.R
\name{import_fonts}
\alias{import_fonts}
\title{import fonts}
\usage{
import_fonts()
}
\value{
loading req fonts
}
\description{
Import fonts Roboto untuk kebutuhan pembuatan judul dan
     subjudul dengan menggunakan fungsi kedata_final
}
\examples{
import_fonts()
}
