% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing0.R
\name{kedata_custom}
\alias{kedata_custom}
\title{Kedata style custome plot finishing}
\usage{
kedata_custom(plotname, title, subtitle, data_source, footer_col = "red", logo)
}
\arguments{
\item{plotname}{ggplot object}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for plot caption}

\item{footer_col}{default to red}

\item{logo}{path to image logo}
}
\value{
ggdraw object
}
\description{
Menempatkan footer dibagian bawah yang terdiri dari logo dan data source
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)

p1 <- ggplot(data = economics, aes(x = pce, y = pop)) +
  geom_line() +
  kedata_theme()

kedata_custom(
  plotname = p1,
  footer_col = "grey",
  logo = "inst/extdata/logo/logokedata.png",
  title = "Lorem Ipsum is simply dummy text",
  subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
  data_source = "www.kedata.online"
)
}

}
