% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing.R
\name{kedata_final}
\alias{kedata_final}
\title{Finishing plot}
\usage{
kedata_final(plotname, title, subtitle)
}
\arguments{
\item{plotname}{ggplot object}

\item{title}{character object}

\item{subtitle}{character object}
}
\value{
ggplot object
}
\description{
Finishing plot
}
\examples{
\dontrun{
library(ggplot2)
# Create usual ggplot object
p1 = ggplot(data = economics, aes(x = pce, y = pop)) +
  geom_line() +
  kedata_theme()
# Drawing left aligned title
p1 <- kedata_final(plotname = p1, title = "Bagian judul ada di sini",
                   subtitle = "Bagian Sub Judul dibawah judul, biasanya lebih panjang dari Judul")
p1
# Saving plot result
ggsave(filename = "tes.png", dpi = 300, height = 6, width = 9)
}

}
