% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing1.R
\name{kedata_final1}
\alias{kedata_final1}
\title{Finishing plot dengan gaya kedata1}
\usage{
kedata_final1(plotname, title, subtitle)
}
\arguments{
\item{plotname}{Sebuah objek yang dibuat dengan menggunakan fungsi ggplot}

\item{title}{Sebuah kalimat berupa karakter untuk judul}

\item{subtitle}{Sebuah kaliamt berupa karakter untuk subjudul}
}
\value{
Objek ggplot
}
\description{
Pertama, membuat judul dan subjudul plot rata dengan tulisan di garis X.
       kedua, membuat logo kedata ada di pojok kanan atas plot yang dibuat.
}
\examples{
\dontrun{
library(ggplot2)
# Create usual ggplot object
p1 = ggplot(data = economics, aes(x = pce, y = pop)) +
  geom_line() +
  kedata_theme()
# Drawing left aligned title
p1 <- kedata_final(plotname = p1, title = "Bagian judul ada di sini",
                   subtitle = "Bagian Sub Judul dibawah judul, biasanya lebih panjang dari Judul")
p1
# Saving plot result
ggsave(filename = "tes.png", dpi = 300, height = 6, width = 9)
}

}
