% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing2.R
\name{kedata_final2}
\alias{kedata_final2}
\title{Kedata style 2}
\usage{
kedata_final2(plotname, title, subtitle, data_source)
}
\arguments{
\item{plotname}{ggplot object}

\item{title}{character}

\item{subtitle}{character}

\item{data_source}{character}
}
\value{

}
\description{
menempatkan footer dibagian bawah yang terdiri dari logo dan data source
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)
p1 = ggplot(data = economics, aes(x = pce, y = pop)) +
  geom_line() +
  kedata_theme()

kedata_final2(plotname = p1,
              title = "Tes Judul di Bagian Ini",
              subtitle = "Tes subtitle di bagian ini lebih panjang",
              data_source = "Sumber Data")
}

}
