% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cor.R
\name{plot_cor}
\alias{plot_cor}
\title{Correlation plot in kedata style}
\usage{
plot_cor(data, title, subtitle, data_source)
}
\arguments{
\item{data}{data frame conatin of all numerci varibale}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for footer text}
}
\value{
ggplot and ggarrange object
}
\description{
Membuat plot korelasi antar variabel
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)

plot_cor(
  data = mtcars,
  title = "Lorem Ipsum is simply dummy text",
  subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
  data_source = "Sumber Data"
)
}
}
