% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dumbbell.R
\name{plot_dumbbell}
\alias{plot_dumbbell}
\title{Dumbbell plot in kedata style}
\usage{
plot_dumbbell(
  data,
  cat,
  x1,
  x2,
  size,
  x_title,
  y_title,
  title,
  subtitle,
  data_source
)
}
\arguments{
\item{data}{data frame}

\item{cat}{character}

\item{x1}{colname 1}

\item{x2}{colname 2}

\item{size}{numerik}

\item{x_title}{character}

\item{y_title}{character}

\item{title}{character}

\item{subtitle}{character}

\item{data_source}{character}
}
\value{
ggplot object
}
\description{
Dumbbell plot in kedata style
}
\examples{
\dontrun{
library(dataplot)

s <- read.csv("https://raw.githubusercontent.com/plotly/datasets/master/school_earnings.csv")
s$School <- factor(s$School, levels = s$School[order(s$Men)])

plot_dumbbell(data = s,
              cat = "School",
              x1 = "Women",
              x2 = "Men",
              size = 3,
              x_title = "Salary",
              y_title = "School",
              title = "Gender earnings disparity",
              subtitle = "Gender earnings disparity from Women VS Men based on University",
              data_source = "www.kedata.online")
}
}
