% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lollipop.R
\name{plot_lollipop}
\alias{plot_lollipop}
\title{Lollipop chart in kedata style}
\usage{
plot_lollipop(data, x, y, title, subtitle, data_source)
}
\arguments{
\item{data}{data frame}

\item{x}{column name}

\item{y}{column name}

\item{title}{character}

\item{subtitle}{character}

\item{data_source}{character}
}
\value{
ggplot objek
}
\description{
Lollipop chart in kedata style
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)

df <- data.frame(
  "brand" = c("Samsung", "Huawei", "Apple", "Xiaomi", "OPPO"),
  "share" = c(10, 30, 20, 35, 5)
)

plot_lollipop(data = df, x = "brand", y = "share",
              title = "Lorem Ipsum is simply dummy text",
              subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
              data_source = "www.kedata.online"
              )
}
}
