% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrg.R
\name{plot_mrg}
\alias{plot_mrg}
\title{Marginal plot on scatter plot}
\usage{
plot_mrg(
  data,
  x,
  y,
  title,
  subtitle,
  data_source,
  mg_type = c("histogram", "boxplot", "density")
)
}
\arguments{
\item{data}{data frame}

\item{x}{character of column name for x axis}

\item{y}{character of column name for y axis}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for footer text}

\item{mg_type}{type of marginal plot}
}
\value{
ggplot ggarrange object
}
\description{
Membuat marginal plot untuk menunjukkan distribusi variabel
             dalam scatter plot
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)

p1 <- plot_mrg(
  data = economics, x = "psavert", y = "uempmed",
  title = "Lorem Ipsum is simply dummy text",
  subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
  data_source = "www.kedata.online",
  mg_type = "histogram"
)

p1
}
}
