% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stackedbar.R
\name{plot_stackedbar}
\alias{plot_stackedbar}
\title{Created stackedbar in kedata style}
\usage{
plot_stackedbar(
  data,
  x,
  y,
  fill,
  title,
  subtitle,
  data_source,
  x_title,
  y_title
)
}
\arguments{
\item{data}{data frame}

\item{x}{colname for x variable}

\item{y}{colname for y variable}

\item{fill}{colname for fill}

\item{title}{character}

\item{subtitle}{character}

\item{data_source}{character}

\item{x_title}{character}

\item{y_title}{character}
}
\value{
ggplot object
}
\description{
Created stackedbar in kedata style
}
\examples{
\dontrun{
library(ggplot2)
library(dataplot)
library(tidyr)
library(dplyr)

df1 <- mpg \%>\%
  count(manufacturer, class)

plot_stackedbar(data = df1,
                x = "manufacturer",
                y = "n",
                fill = "class",
                x_title = "col_x",
                y_title = "col_y",
                title = "Lorem Ipsum is simply dummy text",
                subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
                data_source = "www.kedata.online")
}
}
