% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_treemap.R
\name{plot_treemap}
\alias{plot_treemap}
\title{Create treemap plot}
\usage{
plot_treemap(data, area, fill, subgroup, title, subtitle, data_source)
}
\arguments{
\item{data}{data frame}

\item{area}{numeric data to be plotting}

\item{fill}{categorical data}

\item{subgroup}{categorical data (optional)}

\item{title}{character title}

\item{subtitle}{character subtitle}

\item{data_source}{character data source}
}
\value{
ggplot objek
}
\description{
Create treemap plot
}
\examples{
\dontrun{
library(ggplot2)
library(dataplot)

df <- G20
plot_treemap(data = df,
             area = "gdp_mil_usd",
             fill = "region",
             subgroup = "country",
             title = "Lorem Ipsum is simply dummy text",
             subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
             data_source = "www.kedata.online"
             )

plot_treemap(data = df,
             area = "gdp_mil_usd",
             fill = "region",
             subgroup = "country",
             title = "Lorem Ipsum is simply dummy text",
             subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
             data_source = "www.kedata.online"
             )
}
}
