% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hydro_plotly.R
\name{plotly_hydro}
\alias{plotly_hydro}
\title{Create hydroplot using plotly}
\usage{
plotly_hydro(df, dates, flow, rain, right_label, left_label)
}
\arguments{
\item{df}{data frame/flat data}

\item{dates}{date time format}

\item{flow}{flow data (numeric)}

\item{rain}{rain data (numeric)}

\item{right_label}{text for right label}

\item{left_label}{text for left label}
}
\value{
plotly object
}
\description{
Create hydroplot using plotly
}
\examples{
\dontrun{
library (dataplot)
library(plotly)

rainfall <- dataplot::dataset

plotly_hydro(df = rainfall, dates = rainfall$`Date Time`,
                  flow = rainfall$Discharge,
                  rain = rainfall$`Rainfall depth (mm)`,
                  right_label = "rainfall",
                  left_label = "discharge")
}
}
