% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainfall_data.R
\docType{data}
\name{rainfall}
\alias{rainfall}
\title{Rainfall dataset}
\format{A data frame with 1440 rows and 3 variables:
\describe{
  \item{Date Time}{date time variable}
  \item{Discharge}{discharge data, in double format}
  \item{Rainfall depth (mm)}{rainfall data, in mm double format}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
rainfall
}
\description{
A dataset containing date time, rainfall and discharge data
}
\keyword{datasets}
