% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrdcloud.R
\name{save_wcloud2}
\alias{save_wcloud2}
\title{Saving wordcloud2 object as png}
\usage{
save_wcloud2(graphfile, filename)
}
\arguments{
\item{graphfile}{wordcloud2 object}

\item{filename}{name for saved file}
}
\value{
png files
}
\description{
Saving wordcloud2 object as png
}
\examples{
\dontrun{
library(dataplot)
library(tidytext)
library(tidyr)
library(dplyr)

teks <- dataplot::datatwit
teks <- teks \%>\%
  unnest_tokens(kata, full_text_norm, token = "ngrams", n = 1) \%>\%
  count(kata, sort = TRUE) \%>\%
  head(n = 150)

my_graph <- wordcloud2(teks,
                       size = 2.3,
                       minRotation = -pi/0,
                       maxRotation = -pi/0,
                       rotateRatio = 1,
                       color = "grey",
                       fontFamily = "Roboto")

save_wcloud2(graphfile = my_graph, filename = "coba.png")
}
}
