% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sctr2.R
\name{sctr2}
\alias{sctr2}
\title{Scatter plot 2 variables in kedata style}
\usage{
sctr2(data, x, y, title, subtitle, data_source)
}
\arguments{
\item{data}{data frame}

\item{x}{character of column name for x axis}

\item{y}{character of column name for y axis}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for footer text}
}
\value{
ggdraw object
}
\description{
Membuat scater dengan dua variabel dalam gaya kedata
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)

sctr2(data = economics, x = "psavert", y = "uempmed",
      title = "Lorem Ipsum is simply dummy text",
      subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
      data_source = "www.kedata.online")

ggsave(filename = "tes.png", dpi = 300, height = 6, width = 9)
}

}
