% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctr3.R
\name{sctr3}
\alias{sctr3}
\title{Scatter plot 3 variables in kedata style}
\usage{
sctr3(data, x, y, color, title, subtitle, data_source)
}
\arguments{
\item{data}{Data frame}

\item{x}{character of column name for x axis}

\item{y}{character of column name for y axis}

\item{color}{column name to ploted as point colors}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for footer text}
}
\value{

}
\description{
Scatter plot 3 variables in kedata style
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)
sctr3(data = midwest, x = "area", y = "poptotal", color = "state",
      title = "Lorem Ipsum is simply dummy text",
      subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
      data_source = "www.kedata.online")
}

}
