% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctr4.R
\name{sctr4}
\alias{sctr4}
\title{Scatter plot 4 variables in kedata style}
\usage{
sctr4(data, x, y, color, size, title, subtitle, data_source)
}
\arguments{
\item{data}{Data frame}

\item{x}{character of column name for x axis}

\item{y}{character of column name for y axis}

\item{color}{column name to ploted as point colors}

\item{size}{column name to ploted as point sizes}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for footer text}
}
\value{
ggdraw object
}
\description{
Scatter plot 4 variables in kedata style
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)
sctr4(data = midwest, x = "area", y = "poptotal", color = "state", size = "popdensity",
      title = "Lorem Ipsum is simply dummy text",
      subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
      data_source = "www.kedata.online")
}

}
