% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_plot.R
\name{waffle_plot}
\alias{waffle_plot}
\title{Waffle chart in kedata style}
\usage{
waffle_plot(data, x, y, ndeep, title, subtitle, data_source)
}
\arguments{
\item{data}{data frame}

\item{x}{character of column name for x axis}

\item{y}{character of column name for y axis}

\item{ndeep}{number of waffle}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for footer text}
}
\value{
ggplot object
}
\description{
Waffle chart in kedata style
}
\examples{
\dontrun{
library(ggplot2)
library(dataplot)

df <- structure(list(region = c("Africa", "Asia", "Latin America",
                                "Other", "US-born"),
                     ncases = c(36L, 34L, 56L, 2L, 44L)),
                .Names = c("region","ncases"),
                row.names = c(NA, -5L), class = "data.frame")

waffle_plot(data = df, x = "region", y = "ncases", ndeep = 10,
            title = "Lorem Ipsum is simply dummy text",
            subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
            data_source = "www.kedata.online")
}
}
