% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing2.R
\name{kedata_final2}
\alias{kedata_final2}
\title{Kedata style 2}
\usage{
kedata_final2(plotname, title, subtitle, data_source)
}
\arguments{
\item{plotname}{ggplot object}

\item{title}{character for plot title}

\item{subtitle}{character for plot subtitle}

\item{data_source}{character for plot caption}
}
\value{
ggdraw object
}
\description{
Menempatkan footer dibagian bawah yang terdiri dari logo dan data source
}
\examples{
\dontrun{
library(dataplot)
library(ggplot2)

p1 <- ggplot(data = economics, aes(x = pce, y = pop)) +
  geom_line() +
  kedata_theme()

kedata_final2(
  plotname = p1,
  title = "Lorem Ipsum is simply dummy text",
  subtitle = "Contrary to popular belief, Lorem Ipsum is not simply random text",
  data_source = "Sumber Data"
)
}

}
